refactor: consolidate locale handling logic and remove LocalePathUtils module
